refact: The index.ts file in the CCFG/ directory has been split into multiple files in the src/ folder
