+++ packages/CCFG/package.json
@@ -18,7 +18,7 @@
     "vscode-uri": "^3.0.2"
   },
   "devDependencies": {
-    "@types/node": "^14.17.3",
+    "@types/node": "^14.18.63",
     "@types/vscode": "^1.56.0",
     "@typescript-eslint/eslint-plugin": "^5.28.0",
     "@typescript-eslint/parser": "^5.28.0",
