+++ packages/CCFG/tsconfig.json
@@ -13,6 +13,7 @@
     /* Language and Environment */
     "target": "ES2022",                                  /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
     "lib": ["ES2022"],                                      /* Specify a set of bundled library declaration files that describe the target runtime environment. */
+    "rootDir": "./src",
     "outDir": "./dist",
     // "jsx": "preserve",                                /* Specify what JSX code is generated. */
     // "experimentalDecorators": true,                   /* Enable experimental support for legacy experimental decorators. */
